requirements.txt
flask==3.1.0
flask-login==0.6.3
supabase==2.15.0
python-dotenv==1.0.1
flask-wtf==1.2.1